pyproject.toml
[tool.poetry]
name = "BaggingHSF"
version = "0.0.1"
description = "Training DL segmentation models of HSF using Bagging strategies"
license = "AGPL-3.0-or-later"
readme = "README.rst"
authors = ["Clément POIRET <poiret.clement@outlook.fr>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.9.0"
torchio = "^0.18.45"
pytorch-lightning = "^1.4.0"
pytorch-lightning-bolts = {extras = ["extra"], version = "^0.3.2"}
hydra-core = "^1.1.0"
comet-ml = "^3.13.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
bandit = "^1.7.0"
pylama = "^7.7.1"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
